+++ mne_nirs/channels/tests/test_channels.py
@@ -7,7 +7,7 @@ import os
 from numpy.testing import assert_array_equal
 import mne
 from mne_nirs.channels import list_sources, list_detectors,\
-    drop_sources, drop_detectors
+    drop_sources, drop_detectors, pick_sources, pick_detectors
 
 
 def _get_raw():
@@ -91,3 +91,41 @@ def test_drop_detectors():
     assert 12 in list_detectors(raw_dropped)
     assert 2 not in list_detectors(raw_dropped)
     assert 1 not in list_detectors(raw_dropped)
+
+
+def test_pick_sources():
+    raw = _get_raw()
+
+    sources = list_sources(raw)
+    assert_array_equal(sources, range(1, 9))
+
+    raw_picked = pick_sources(raw.copy(), 3)
+    assert 3 in list_sources(raw_picked)
+    assert 2 not in list_sources(raw_picked)
+    assert 1 not in list_sources(raw_picked)
+
+    raw_picked = pick_sources(raw.copy(), [7, 2])
+    assert 7 in list_sources(raw_picked)
+    assert 2 in list_sources(raw_picked)
+    assert 1 not in list_sources(raw_picked)
+    assert 11 not in list_sources(raw_picked)
+    assert 9 not in list_sources(raw_picked)
+
+
+def test_pick_detectors():
+    raw = _get_raw()
+
+    sources = list_detectors(raw)
+    assert_array_equal(sources, range(1, 17))
+
+    raw_picked = pick_detectors(raw.copy(), 3)
+    assert 3 in list_detectors(raw_picked)
+    assert 2 not in list_detectors(raw_picked)
+    assert 1 not in list_detectors(raw_picked)
+
+    raw_picked = pick_detectors(raw.copy(), [7, 2])
+    assert 7 in list_detectors(raw_picked)
+    assert 2 in list_detectors(raw_picked)
+    assert 1 not in list_detectors(raw_picked)
+    assert 11 not in list_detectors(raw_picked)
+    assert 9 not in list_detectors(raw_picked)